requirements.txt
nose==1.1.2
parsedatetime==0.8.7
python-dateutil==2.1
